+++ tests/testdata/endpoints.json
@@ -30,6 +30,6 @@
     "observation_tracktimerange": "/api/v1.0/observation/tracktimerange/string:platform_id.json",
     "observation_track": "/api/v1.0/observation/track/platform_type=none.json",
     "observation_point": "/api/v1.0/observation/point/platform_type=none.json",
-    "observation_meta": "/api/v1.0/observation/meta.json",
+    "observation_meta": "/api/v1.0/observation/meta/platform/123.json",
     "observation_variables": "/api/v1.0/observation/variables/station=0.json"
 }
